+++ src/command_modules/azure-cli-sql/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 Unreleased
 +++++++++++++++++++
 * Adding support for SQL Transparent Data Encryption (TDE) and TDE with Bring Your Own Key
+* Added az sql db list-deleted command and az sql db restore --deleted-time parameter, allowing the ability to find and restore deleted databases.
 
 2.0.12 (2017-09-22)
 +++++++++++++++++++